add options
